ar format: Handle the case where ar lists files with single-digit dates
A function for finding the start of the archive member's name used to
consider two spaces in a row to mean two "fields" in the formatted date
string produced by ar.  This led to pathnames errantly containing the
year the file was created.  This commit considers double spaces to mean
the same thing that single spaces mean.