Fix MIME detection logic from #490
A blooper has been made there:

* if ENABLE_MIME is set, the intention was to try, in order: magic,
  content, filename; but it was made filename, content, magic (which
  was the same as before the changes);
* if ENABLE_MIME is not set, the intention was to try, in order:
  filename, content, magic; but it has been made magic, content, magic
  (notice the duplicate, and the missing "filename").

This probably doesn't change much in the wild as magic is gonna work
most of the time, but it's especially problematic that the non-libmagic
case doesn't have the filename test.

Anyway, fix this so the code is consistent, and we retain the behavior
for the non-libmagic case, and have the new expected one for the
libmagic case.