rar: Simplify and merge duplicate code
Merge version matching for rar and unrar.  This is ever so slightly
slower in theory for unrar 5.x because there's an additional
unnecessary `sscanf()` call, but it's not actually gonna matter, and
saves quite a bit of duplicated logic.

This also makes the matching a bit more safe by verifying `sscanf()`
actually worked, and avoids using a uninitialized `version` value due
to an unexpected input.

Finally, this makes the parsing a bit more strict by requiring a
version number after the `RAR` and `UNRAR` line prefixes -- leading not
to using the uninitialized version variable.
This will be required by the upcoming unrar-free support as it reports
a `RAR archive ...` line that would have matched the `RAR` version
check.