rar: Fix asking password for encrypted archives with RAR7
RAR 7.0 outputs a different error when password is missing:

>
> Incorrect password for foo.rar
> Incorrect password for foo.rar
> Cannot find volume foo.r00

Whereas RAR 6.21 doesn't output the last 2 lines.

The error about volumes confuses the parser thinking there's a volume
issue, which hides the incorrect password one.

Note that RAR 7.0 outputs this *most of the time*, but once in a while
it'll also skip the last 2 lines.  Whether it is a race condition or
not, it at least happens often enough with RAR 7.0 to consider this its
normal behavior and handle it.

Fixes #526.